consolidate vote application
