+++ core/tests_vote_engagement.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 
 from .models import Community, EngagementEvent, Post
-from .services.votes import apply_vote
+from .votes import apply_vote
 
 
 class VoteEngagementTests(TestCase):
